Move the GradleEnteprise features to a more logical place
Now we support Gradle and Maven, having these in the Maven package doesn't make sense.

This PR moves them to the parent package.

Closes #1711
